pyproject.toml
[project]
name = "gemai"
version = "0.1.0"
description = "GenAI Code comment"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.30.0",
    "langgraph>=0.6.5",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
